Facilitate full disk encryption
This does not add full disk encryption, since that is outside the
scope of managing grub. What this does is configuring of grub, with
user or otherwise provided variable data, so it is able to boot
fully encrypted disks.

There are some requirements for this to work.

1. The grub_boot_device has to be defined, which was always the case.
2. All values in the boot_encryption dictionary has to be defined and
   boot_encryption.enabled has to be true (this enables the feature).
3. The dictionary values of:
   - boot_encryption.cryptdevice
   - boot_encryption.rootdevice
   has to be set according to documentation and example files.
4. The full disk encryption setup has to be LVM on LUKS. See:
   https://wiki.archlinux.org/index.php/Dm-crypt/Encrypting_an_entire_system#LVM_on_LUKS

There may be other limitations or requirements, but those are the
most important. See the main target for this project for more info:
https://github.com/AsavarTzeth/ansible-archlinux-desktop.git