Remove mkinitcpio hook
This is due to conflicts with mkinitcpio role in main project.

A simple solution is to sacrifice one for the other. The reasoning is
that changes to /etc/mkinitcpio.conf are far more likely than
/etc/default/grub. At least this is true with the current state of
the main project, where grub changes would only be made if
cryptdevice or root kernel parameters changed.